requirements.txt
fastapi~=0.115.5
uvicorn~=0.32.0
pydantic~=2.9.2
psycopg2~=2.9.10
PyJWT~=2.9.0
python-dotenv~=1.0.1